refactor: move PodcastSearchResultList to features directory
